IMPALA-12371: Add better cardinality estimation for Iceberg V2 tables with deletes
Currently IcebergDeleteNode's cardinality is the same as the LHS's
cardinality, i.e. we don't take the RHS into account. The RHS contains
the position delete records, so it is a fair assumption that all records
at RHS remove a record from LHS (duplicated delete records should be
extremely rare).

If there are conjuncts on the Iceberg table we can assume that they have
the same selectivity on the data records and on the delete records.

With the above assumptions this change updates the cardinality of the
IcebergDeleteNode with basically the following formula:

 Card(IcebergDeleteNode) = Card(LHS) - Selectivity(LHS) * Card(RHS);

Please note that left side's cardinality already takes the selectivity
into account.

To deal with edge cases when there are lots of duplicated delete
records (shouldn't happen in normal usage), we return at least 1
cardinality, so the actual formula is:

 Card(IcebergDeleteNode) =
   Max(
     1,
     Card(LHS) - Selectivity(LHS) * Card(RHS)
   );

Testing:
 * updated the planner tests

Change-Id: I988dc8d7e1074932c460b3702d3381341e5b23c5
Reviewed-on: http://gerrit.cloudera.org:8080/20460
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>