IMPALA-12313: (part 1) Refactor modify statements
This change refactors the classes and methods that implement
modify statements like DELETE and UPDATE. ModifyStmt, DeleteStmt,
UpdateStmt are created during parsing and contain information about
the statement: FROM clause, WHERE clause, target table, etc.

The logic that actually implements these operations is dependent
on the type of the target table. Therefore during analysis, after
the target table is resolved, we create the *Impl object (e.g.
IcebergDeleteImpl, KuduUpdateImpl) that implements the logic. The
impl object is in charge of creating the source statement of the
operation, doing the necessary rewrites/masking, and also creating
the data sink.

Testing:
 * N/A: no new functionality / bug fix

Change-Id: If15f64944f2e23064b7112ad5930abc775dd65ec
Reviewed-on: http://gerrit.cloudera.org:8080/20477
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>